// src/interfaces/types.ts
export type Items = {
  icon: string
  label: string
  type: string
  pageUrl: string
}

export type MenuItems = {
  icon: string
  label: string
  pageUrl: string
}

export type Absences = {
  users: string
  category: string
  periode: string
  soldeActuel: number
  soldePris: number
  soldeFutur: number
}

export type Users = {
  name: string
}